+++ src/main/java/nextstep/qna/domain/course/Course.java
@@ -1,11 +1,16 @@
-package nextstep.qna.domain;
+package nextstep.qna.domain.course;
+
+import nextstep.qna.domain.session.Session;
 
 import java.util.List;
 
 public class Course {
 
+    private Long id;
+
     private String title;
 
     private List<Session> sessions;
+
 
 }
