+++ src/main/java/nextstep/qna/domain/enums/ContentType.java
@@ -1,4 +1,4 @@
-package nextstep.qna.domain;
+package nextstep.qna.domain.enums;
 
 public enum ContentType {
     QUESTION, ANSWER;
