+++ src/main/java/nextstep/qna/domain/question/Question.java
@@ -1,7 +1,11 @@
-package nextstep.qna.domain;
+package nextstep.qna.domain.question;
 
 import lombok.Getter;
 import nextstep.qna.CannotDeleteException;
+import nextstep.qna.domain.answer.Answer;
+import nextstep.qna.domain.answer.Answers;
+import nextstep.qna.domain.enums.ContentType;
+import nextstep.qna.domain.history.DeleteHistory;
 import nextstep.users.domain.NsUser;
 
 import java.time.LocalDateTime;
