+++ src/main/java/nextstep/qna/repository/AnswerRepository.java
@@ -1,4 +1,6 @@
-package nextstep.qna.domain;
+package nextstep.qna.repository;
+
+import nextstep.qna.domain.answer.Answer;
 
 import java.util.List;
