+++ src/main/java/nextstep/qna/repository/DeleteHistoryRepository.java
@@ -1,4 +1,6 @@
-package nextstep.qna.domain;
+package nextstep.qna.repository;
+
+import nextstep.qna.domain.history.DeleteHistory;
 
 import java.util.List;
