+++ src/main/java/nextstep/qna/service/DeleteHistoryService.java
@@ -1,7 +1,7 @@
 package nextstep.qna.service;
 
-import nextstep.qna.domain.DeleteHistory;
-import nextstep.qna.domain.DeleteHistoryRepository;
+import nextstep.qna.domain.history.DeleteHistory;
+import nextstep.qna.repository.DeleteHistoryRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
