+++ src/test/java/nextstep/qna/domain/QuestionTest.java
@@ -2,6 +2,7 @@ package nextstep.qna.domain;
 
 import nextstep.qna.CannotDeleteException;
 import nextstep.qna.domain.fixture.QuestionFixture;
+import nextstep.qna.domain.history.DeleteHistory;
 import nextstep.users.domain.NsUserFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
